Updated delete to return meaningful data if possible
