+++ response_data/projects.py
@@ -1,0 +1,1 @@
+delete_project = {}
